Optionally stream the stderr and stdout of run_command
This feature would enable us to monitor long running tasks, such as an
analysis over a really large codebase, etc. That could take 10-30 hours,
and sometimes the process hangs as well.
In CI systems, collecting logs is really important to let us track down
the related issues. It does not matter if the log is verbose.